src/requirements.txt
oracledb
python-dotenv
